+++ packages/utilities/src/m3/index.ts
@@ -1,0 +1,6 @@
+export * from "./diagrams/Mermaid-generator.js"
+export * from "./diagrams/PlantUML-generator.js"
+export * from "./ecore/importer.js"
+export * from "./ecore/types.js"
+export * from "./schema-generator.js"
+export {asText} from "./textual-syntax.js"